requirements.txt
asgiref==3.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
lxml==6.0.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
playwright==1.54.0
pluggy==1.6.0
pycparser==2.22
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
